Fix dependency issue between initial classes
